Added event types dropdown to ManageEvent.js and implemented /eventTypes endpoint in EventController.java
In ManageEvent.js, added a useEffect hook to fetch event types from the /eventTypes API and store them in a state. Used this state to populate a dropdown menu for event types.

In EventController.java, implemented a new /eventTypes endpoint that returns a list of all event types. This endpoint is used by ManageEvent.js to populate the event types dropdown.